replace f64 precision with f32

+++ src/args.rs
@@ -6,9 +6,9 @@ pub struct Cli {
     #[arg(short, long)]
     pub fixed_ball_spacing: bool,
     #[arg(short, long, default_value_t = 0.0)]
-    pub white_offset: f64,
+    pub white_offset: f32,
     #[arg(short, long, default_value_t = 2.0)]
-    pub initial_velocity: f64,
+    pub initial_velocity: f32,
     #[arg(short, long, default_value_t = String::from("./output.xyz"))]
     pub simulation_output_path: String,
     #[arg(short, long, default_value_t = String::from("./times_output.txt"))]
