+++ src/io.rs
@@ -118,7 +118,7 @@ pub fn output_snapshot(file: &mut File, table: &Table) -> Result<()> {
     Ok(())
 }
 
-pub fn output_event_times(path: &str, events_times_list: &[(f64, usize)]) -> Result<()> {
+pub fn output_event_times(path: &str, events_times_list: &[(f32, usize)]) -> Result<()> {
     let mut output = File::create(path)?;
 
     writeln!(output, "{}", events_times_list.len())?;
